Switched to Visual Studio Enterprise 2022 (some changes were made in the source code). Added social media links: Twitter, LinkedIn, Facebook, and Instagram. Added shortcuts to GitHub repository's Issues, Discussions, and Wiki.

+++ UnquoteHTML.cpp
@@ -330,7 +330,7 @@ const std::string UnquoteHTML(const std::string& InBuffer)
 	bool ProcessedChar, GotCharCode;
 	MatchState = NoMatch;
 	std::string OutBuffer;
-	for (int index = 0; index < InBuffer.length(); index++)
+	for (size_t index = 0; index < InBuffer.length(); index++)
 	{
 		const unsigned char ThisCh = InBuffer[index];
 		ProcessedChar = false; /* to begin with */
